requirements.txt
streamlit==1.28.0
pandas==2.0.3
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.4.2
python-docx==0.8.11
plotly==5.15.0
google-cloud-aiplatform==1.38.1
google-cloud-vision==3.4.4
protobuf==4.24.4